moves web3 hooks to proper place
