--- src/hooks/index.ts
@@ -1,5 +1,0 @@
-import useETHGasPrice from './useETHGasPrice';
-import useInitContracts from './useInitContracts';
-import useIsDefaultChain from './useIsDefaultChain';
-
-export {useETHGasPrice, useInitContracts, useIsDefaultChain};
